feat: Update `CorsetValidator` with Expansion Options (#759)
* Update `CorsetValidator` with Expansion Options

This updates the `CorsetValidator` to support various configuration
options used by `corset`.  Specifically, for field arithmetic and
constraint expansion.  A better set of defaults is used which means
testing now reflects more closely what the prover sees.  However, we
stop short of going all the way to the lowest level because this has a
large impact on performance.  Nevertheless, the option to do this is
present in the updated `CorsetValidator` and can be enabled easily
enough.

* Configure `CorsetValidator` via `CORSET_FLAGS`

This allows the `CorsetValidator` to be configured via an environment
variable `CORSET_FLAGS`.  This allows, for example, different levels of
expansion to be specified on the CI pipeline, or when running locally.

In addition, this updates the `CorsetValidator` to allow different
amounts of expansion to be specified.  Previously, it was "all or
nothing".  Now, upto four levels of expansion can be specified.

* Update GitHub Action

This updates the github action to override the default configuration for
`corset`.  Specifically, it uses slightly less expansion than the
default.

* Fixes for `spotless`

* Update `README`

This updates the `README` with some information on how to disable corset
expansion to get faster testing.

* Fix spotless (again)