feat: update gradle test config to show better output (#1021)
* Update gradle test config to show better output

This updates the gradle test configuration to generate better output
during testing.  In particular, so we can see why a particular test
failed rather than just "FAILED".  This will be helpful for both Rust
`corset` and `go-corset`.

* Prohibit standard streams on passing tests

This disables standard streams from being printed for passing tests.
That should just reduce the clutter when a test is failing.