fix: use self hosted runner for blockchain tests (#1620)
* Update Gradle to v8.11.1

This updates the gradle wrapper (i.e. gradlew) to version 8.11.1.  This
is currently the latest version of gradle.

* Reduce `REFERENCE_TESTS_PARALLELISM`

This number is set too high for the given runner.  I've reduced it to
just `1` for now to try and get things working again.  Subsequently we
can try incrementing this to find the sweet spot.

* Specify `GOMEMLIMIT`

* Tweak memory settings

This tweaks the memory settings to facilitate better performance on the
CI server.  In fact, it lowers the amount of memory available.  This is
because we don't want Java and Go competing for the same amount of
memory.  Furthermore, we now know that all of the tests can be completed
on a machine with just 32Gb RAM.