Feat/blockdata unique gas limit constant (#1699)
* fix `gradle build`

This ensures that, when running `gradle build`, it also arises the
`buildTracer` is run beforehand.  This is necessary to ensure that all
`Trace.java` files (and related) are generated.

* force trace files to build beforehand

This attempts to force the `Trace.java` files to be built before
anything else important happens.

* fix(tracejava): replay tests

* feat(tracejava): reintegrate blockhash

* feat(blockdata): add linea and ethereum tracer build task

* fix(blockdata): remove getter

* feat(blockdata): add eth trace task for ref tests and gaslimit test fix

* feat(blockdata): blockhash generation trace reintegration

* feat(blockdata): rebase on 1695 branch

* feat(blockdata): update constraints

* feat(blockdata): remove trace java file

* Enable conditional build

This configures the build process to be conditional on a system-wide
property called "blockchain".  This property can be specified on the
command-line with e.g. "-Dblockchain=Ethereum" to set the target chain
to Ethereum.  By default, the blockchain is assumed to be Linea.
Therefore, its only when running the reference tests that this property
needs to be set.  Furthermore, if the reference tests are run without
setting this property ... then an error is reported and the build fails.

---------

Co-authored-by: DavePearce <dave01001110@gmail.com>