feat: go-corset front end integration (#1689)
* Use `go-corset` front-end

This puts through some changes to use the new `go-corset` front end.

1) This updates the gradle-tests action to use the `go-corset` front-end
instead of the original.  Provided this works, other actions will need
to be updated accordingly.

2) This updates the version of `go-corset` used in `linea-tracer`.  This is
necessary because only the latest version of `go-corset` is capable of
compiling the constraints.

3) This makes relevant changes to ensure go-corset can compile the
zkevm.bin file as necessary for the reference tests.

4) This updates the corset version to incorporate a minor fix relating to
coalescing of i8 and byte columns.