fix(RewardsStreamerMP): ensure `registerVault` reverts in emergency mode
Originally, I just wanted to simplify the certora rule that,
when emergency mode is enabled, only a few selected functions can be
called.

Instead of listing all the view function as "allowed", I've moved to
using CVLs `isView` flag on the function under verification.

This however uncovered a violation where
`RewardsStreamerMP.registerVault` is allowed to be called in emergency
mode.

In theory there's no harm in registering a vault when the system is in
emergency mode, but semantically it doesn't really make sense.

`registerVault` has been accidentally added to `isViewFunction()`.
This commit fixes this by adding `onlyNotEmergencyMode` modifier to
`registerVault()`.