add explicit types for blockdata gas limit (#609)
This adds explicit types to the constants `GAS_LIMIT_MINIMUM` and `GAS_LIMIT_MAXIMUM` for all three currently supported changes: `Linea`, `Ethereum` and `OldLinea`.  This change is necessary to ensure that, when the `Trace.java` file is generated, the constant always has the same type.  Since these are given the type `i64` here, they will always generate a Java constant of type `BigInteger`.