feat: utilise parameterised zkevmbin (#1890)
The ChainConfig encapsulates key information about the chain being
traced.  Specifically, it allows us to set the upper/lower gas limit
bounds, and enable/disable the block gas limit.  It also encapsulates
the ChainID for reasons of simplicity.

1. This updates the replay and reference test frameworks to use the
ChainConfig.

2. This updates the GoCorsetValidator to configure the chain parameters
when validating a trace.

3. This updates the CI config to avoid using e.g. "-Dblockchain=Ethereum"
as this is no longer necessary.  This also fixes an issue related to the
BlockchainReferenceTests whereby the ChainConfig was not being set
whilst creating ZkTracer.

4. This also puts the LineaL1L2BridgeSharedConfiguration as a field of
ChainConfig.  This helps to reduce the number of constructors needed for
ZkTracer, and clearify when testing environments are used versus
non-testing environments.

Signed-off-by: F Bojarski <ceciestunepoubelle@protonmail.ch>

---------

Signed-off-by: F Bojarski <ceciestunepoubelle@protonmail.ch>
Co-authored-by: F Bojarski <ceciestunepoubelle@protonmail.ch>