shanghai constraints and testing (#1993)
* update shanghai constraints

Signed-off-by: F Bojarski <ceciestunepoubelle@protonmail.ch>

* Paris HF

Signed-off-by: F Bojarski <ceciestunepoubelle@protonmail.ch>

* fork dependant line counting for txn data

Signed-off-by: F Bojarski <ceciestunepoubelle@protonmail.ch>

* fix ToyExecutionTools + ChainConfig

Previously, the ToyExecutionTools were always using a fixed ChainConfig
to initialise CorsetValidator.  This meant that the ChainConfig used by
CorsetValidator and ZkTracer could be out of sync (e.g. one for London
and the other for Shanghai).  This is fixed now.

* improve naming for test lt files

This now includes the fork in the filename for trace (i.e. lt) files
generated from the unit tests.  Just helps to clarify which was which.

---------

Signed-off-by: F Bojarski <ceciestunepoubelle@protonmail.ch>
Co-authored-by: DavePearce <dave01001110@gmail.com>
Co-authored-by: amkCha <29160563+amkCha@users.noreply.github.com>