feat: fix state tests (#2025)
This puts in place an optional filter on tests to run for the general
state tests.  This is available when running the action directly via
github, exactly as it already is for the blockchain reference tests.

In addition, this tweaks a few parameters for both the state and
blockchain reference tests.

Finally, this moves the blockchain reference tests over to the 
self-hosted runners.