feat: run unit tests per fork (#2035)
* feat: testInfoWithChainConfig in builder toyExecEnv

* feat: set replay to London fork

* feat: multiblock testinfo

* feat: refacto Bytecode dynamic loadOpcodes

* fix outstanding compilation errors

This fixes the outstanding compilation errors, such that the trace now
compiles and passes at least one test.  However, I believe there are
still problems related to the use of testInfo in static variable
initialisers.  Specifically, we will see that it is null in such
situations, leading to some kind of exception.

* feat: parameterize unit tests

---------

Co-authored-by: DavePearce <dave01001110@gmail.com>