initial refactoring for go-corset v1.1.15
This puts through an initial (but not yet complete) refactoring for
go-corset v1.1.15.  The main differences are around: (1) the
organisation of packages; (2) a more complex compilation pipeline which
supports mixed assembly and corset modules; (3) merging of the HIR and
MIR levels together; (4) changes in the way interleaving assignments are
handled.