fix: trace limits
This fixes two distinct issues with trace limits.  The first is around
the names of modules for the "limitless type proofs".  Previously, these
were called e.g. ":u128".  Now they are called e.g. "u128".

The other issue is around modules with non-trivial length multiplier.
Specifically, a module "m1" with a length multiplier of 3 has the name
"m1×3".  Its not clear to me what the best process for resolving this
is, so I leave it for others.