feat: refactor `OpCodes` API (#2204)
* refactor OpCodes API to be fork agnostic

Previously, the OpCodes API was implemented via a number of static
methods and, once a given fork was loaded, it was not possible to load
another fork (doing so would cause a runtime exception).

Therefore, this refactors the OpCodes API to remove entirely the static
API and, instead, replace it with an API based around OpCodes
*instances*.  That is, we have an instance of OpCodes for each fork
under consideration.  This means that, in situations where previously we
could look up OpCodeData information statically, we must now be able to
access a relevant instance of OpCodes.  As such, quite a lot of
refactoring was required to make this possible.

The interface now promoters OpCodeData over OpCode as the primary source
of useful information.  The OpCode class is now largely relegated to
being simply a "mnemonic" with essentially no actual functionality.  In
most cases, this works out well because OpCodeData already had a
mnemonic accessor, meaning whereever we have an instance of OpCodeData
we also have an instance of OpCode.

* fix OpCodeData.equals => OpCode.equals

* fix StpCall

The StpCall class was incorrectly using an OpCodeData field instead of
an OpCode field.  Specifically, this was being used within its equals
and hashCode methods.

* use @Getter instead of getOpcodes()

* fix: BytecodeCompiler.op()

This function previously threw an exception if the opcode provided was
invalid for the given fork.

* Use @BeforeAll in conjunction with @BeforeEach

The issue with using @BeforeEach to set the fork and load the opcodes is
that this is not guaranteed to run before methods used as MethodSource's
for parameterised test.  However, @BeforeAll is guaranteed to run before
any test, including any MethodSource methods.  Therefore, splitting up
the process of initialising the TracerTestBase gives the best solution.