chore(StakeManager): clean up `StakeVault` integrity checks
This commit removes the unnecessary check in `StakeManager.registerVault`
that the calling `StakeVault` is indeed referencing the correct `StakeManager` instance.

This was pointed out in: https://github.com/vacp2p/staking-reward-streamer/issues/245

It also moves the vault owner check in `StakeVault.migrateToVault` into the `StakeVault` as
the `StakeVault` is trusted via `trustedCodehash` modifier.

Closes https://github.com/vacp2p/staking-reward-streamer/issues/245