fix(StakeVault): ensure StakeManager accouning is done before withdraing
There was a bug in `withdraw()`, which causes inconsistent state in `StakeManager` with users
withdraw their funds, as nothing would ensure that the stake manager's internal accounting would be updated properly.

In fact, withdrawing staked funds should only be possible when a user has called `leave()` first and `lockUntil` has expired.

This commit adds a `hasLeft` flag, which is set when `leave()` was called.
It also adds a Certora rule that checks which functions are still allowed to be called after leave mode has been enabled.