feat: refactor GAS / ADD tracers for asm implementations (#2234)
* refactor GAS / ADD tracers for asm implementations

This refactors both of these modules to account for their corresponding
zkasm implementations.  For GAS, we have an entirely new implemetnation
replacing the original Corset constraints.  For ADD, we are dropping
legacy support in the LONDON fork for the non-assembly implementation.

* update linea-constraints commit to latest master

* tweak AddOperation and Gas/GasOperation

This applies some simple tweaks to these modules to tidy them up a
little.  For example, removing the now unnecessary wcp argument for Gas
and GasOperation.