feat: improve tracefile naming for parametric tests (#2236)
* preliminary testing

This threads through the testInfo object into the ExecutionEnvironment
such it can be used as part of the test name.  For parametric tests,
this contains information about the parameters used.

At this stage, this only prints out the information for demonstration
purposes.  It does not use it to construct a filename (yet).

* enable tracefile naming from (processed) TestInfo

This puts in place code necessary for processing the TestInfo
information generated with a given test.  However, this is a fairly
significant and outstanding issue, namely that the static field within
the TracerTestBase class cannot be relied upon in a concurrent setting.

* remove TestInfoWithChainConfig

This removes the class TestInfoWithChainConfig and, as a result, applies
a large refactoring.  The root cause for this was the presence of the
TestInfo information being stored in TracerTestBase.  It cannot sensibly
be stored at this point when tests are running concurrently.  Instead,
we must obtain the TestInfo object from the test method itself as an
additional parameter.  This issue was impacting the name generation for
parameterised tests based on TestInfo data.  Specifically, the
information was often inaccurate as it was involved in a race condition.

Refactoring the tests to include the TestInfo parameter meant the
benefit of the TestInfoWithChainConfig class was essentially
eliminated.  In particular, to continue using it would require
constructing a TestInfoWithChainConfig instance within each test.
Basically, its easier to remove the class entirely than actually do
that.  In some cases, this leads to a more precise flow of information,
whereby unnecessary TestInfo data is not provided to functions which
don't require it.

* include parameter names in filename

This includes the parameter names in the filename.  However, for reasons
unknown (probably lombok related), the parameter names have been changed
from their originals.  So this limits the benefits of this, but it still
helps actually a bit.

* apply spotless

* stop legacy tracefile naming for replays

This removes the use of legacy tracefile naming from the replay tests.
This just ensures we are now using a consistent naming scheme across all
tests.