feat: support file system caching of trace files (#2250)
* preliminary implementation of tracefile caching

This puts in place a preliminary (and completely untested)
implementation of tracefile caching.  In particular, this attempts to
read the existing tracefile (if it exists) and check the embedded
version number.  If all of these checks pass, it will reuse the trace
file.

* rework tracefile parser

This reworks the tracefile parsing logic to be more robust.  For
example, if we want to extract other metadata attributes.  Also, it has
now been tested on some real trace files.

* Add option to CLI

* remove required as default is set.

* remove LtTraceFile header check

The header check is unnecessary as the releaseVersion is already
embedded in the filename.  Therefore, if the file exists, we know its
from the right releaseVersion.

---------

Co-authored-by: FlorianHuc <florian.huc@gmail.com>