feat(Karma): add transfer whitelisting capabilities
This adds functions to allow whitelisted accounts to transfer their
Karma. The tranferable karma is exclusively the one that lives in the
Karma contract. Karma coming from reward distributors are not transferable
and need to be converted first.

The transfer function also takes the `slashAmount` into account. Whitelisted
accounts cannot send more than what is transferable.

The commit adds a few tests and also adjusts one of the certora rules accordingly.