feat: integrate assembly shf and exp modules (#2283)
* update Exp and ExpOperation

This updates theses classes to target the assembly implementation of the
EXP module.  Since this does the trace generation itself, we no longer
need to support preprocessing for the perspectives.

* update ShfOperation for assembly implementation

This updates the ShfOperation for the assembly implementation which, as
it does most of the trace generation, means there is less to do within
ShfOperation.

NOTE: this also removes the ShfRt table, as this is not used by the
assembly implementation.

* fix spotless

* minor fixes & update go-corset

This puts through some minor fixes, sets previously disabled EXP tests
to run on nightly and updates go-corset to the required version.

* update EXP asm implementation

This puts in place an update for the EXP asm implementation to account
for the case where CDS < EBS.  In such case, the bytes from CDS upto EBS
need to be filled with zero.

NOTE: this implementation has now been tested against *all* tests in
ExpTest, including those tagged as "nightly".

* tidying up to address review comments

This tidies up a few things, such as using the new switch statement
syntax, and the existing OpCode.unsignedByteValue() method.

* update `linea-constraints` to master

This updates `linea-constraints` to the latest commit of master.