fix: missing equals and hashcode for `ExpCall` and `GasParameters` (#2296)
* add toString() methods to ExpCall implementations

This adds some useful methods for debugging to the various ExpCall
implementations.

* add equals() and hashCode() to ExpCall impls

This adds equals() and hashCode() methods to the two ExpCall
implementations.

* add equals() and hashCode() methods to GasParams

The GasParameters class is used within GasOperation and, hence, should
implement equals and hashCode().  Unfortunately, it previously did not.

* add abstract methods to OobCall

The purpose of this is to make sure any subclasses (now or in the
future) have these methods.

* fix spotless

---------

Co-authored-by: François Bojarski <54240434+letypequividelespoubelles@users.noreply.github.com>