prover: bump go-corset and constraints (#1317)
* bump go-corset to v1.1.15

* update go-corset to v1.1.15

* bump constraints to beta-v3.1-rc7

* initial refactoring for go-corset v1.1.15

This puts through an initial (but not yet complete) refactoring for
go-corset v1.1.15.  The main differences are around: (1) the
organisation of packages; (2) a more complex compilation pipeline which
supports mixed assembly and corset modules; (3) merging of the HIR and
MIR levels together; (4) changes in the way interleaving assignments are
handled.

* finalise and update to go-corset to v1.1.16

The update to v1.1.16 was required to give an accessor method for
air.Constraint in order to access the actual (underlying) constraint.

* fix: trace limits

This fixes two distinct issues with trace limits.  The first is around
the names of modules for the "limitless type proofs".  Previously, these
were called e.g. ":u128".  Now they are called e.g. "u128".

The other issue is around modules with non-trivial length multiplier.
Specifically, a module "m1" with a length multiplier of 3 has the name
"m1×3".  Its not clear to me what the best process for resolving this
is, so I leave it for others.

* update to go-corset v1.1.19

This integrates the prover with go-corset v1.1.19, which includes a
number of changes related to field agnosticity and ZK assembly.

* update gitattributes

This updates the git attributes to inform git that the zkevm.bin file
is binary, and should have its line endings changed.

* update to go-corset v1.1.20

This is a precaution as go-corset version v1.1.19 contains a race
condition.

* update for assembly modules exp / shf

* update `linea-constraints` to `9cad732`

This also updates the `zkevm.bin` file to the compiled
`zkevm_london.bin`

* fix broken limit config

* update constraints to beta-v4.0-rc10 (edbf0c6)

* replace fmt.Printf with logrus.Infof for module size logging in scanColumns

* update sepolia config for beta v4.0

* update devnet config

* update type TraceLimits for beta v4.0

* add trace limits for new modules in beta v4.0

* fix e2e error: remove unused column handles from GetAffinities function

* add scaling for new beta 4.0 internal modules in TracesLimits

* remove unused SHF_REFERENCE_TABLE from integration benchmark and full config files

* add new discovery advice entries and update existing column references in limitless.go

* remove unused column handle from GetAffinities function

* update trace limits for beta v4.0

* update beta v4.0 module limits

* add default trace limits for Beta v4.0

* add missing discovery advice entries for beta v4.0

---------

Signed-off-by: Leo Jeong <dreamerty@postech.ac.kr>
Co-authored-by: DavePearce <dave01001110@gmail.com>