fix HubShomeiTests.opcodeProvider() (#2357)
This method had a bug whereby it was intending to test both the SSTORE
and SLOAD cases but, in fact, was only testing the SSTORE case.  When
the TestInfo parameter was added as part of a previous PR, this bug was
exposed because it meant JUnit was trying to fill the TestInfo class
with an SLOAD opcode --- which does not work.