fix: tidy up after replay tests PR (#2471)
This tidies two things up after the replay tests PR.  Firstly, the
`build.gradle` file for the reference-tests is updated to use
`JUNIT_TESTS_PARALLELISM` to match the github workflow; secondly, the
disable replay test in BlockhashTest is removed, since it was a
duplicate.  Specifically, the replay tests for BlockHash are now in
`pragueReplayTests/BlockHashTests`.