feat: support in memory trace compression (#2504)
* support in memory trace generation

This alters the way in which trace generation occurs, in that it traces
are always generated fully in memory before being written to disk.  This
allows various possibilities.  For example, we can compress the data
prior to writing to disk.  Or, we can use an alternative format which
requires holistic information of the entire trace before being written
to disk (e.g. it uses some kind of heap, etc).

* add plugin cli option for trace compression

Trace compression should be enabled by default, but employs a feature
flag which can be used to disable it.

* update go-corset to v1.1.29

This updates go-corset to a version which is compatible with the trace
compression code used by this PR.

* add logging and minor comments

* fix out-of-memory issue

There was an issue due to the maximum size of a Java byte array (~2G).
Specifically, the entire trace was being put into a single byte array,
and this was causing an overflow.  Instead, we just write directly using
output streams.