feat: find registers using source-level names
This adds support for finding registers based on their source-level
name.  For example, we might have a column written in the Corset lisp
file called "DATA" that we want to access.  For whatever reason, this
column might be mapped to a raw register called e.g. "DATA_xor_HASH".
Whilst we could use the raw register name directly, it is subject to
change as the constraints change.  Instead, we want to be able to use
just "DATA" and for it to figure out the raw name of "DATA_xor_HASH" for
us.  This is what the Arithmetisation.LimbsOf() function now does.